learnmore/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
Django==5.2.1
django-allauth==65.8.0
djangorestframework==3.16.0
idna==3.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
